% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_annotation.R
\name{annotate_clusters}
\alias{annotate_clusters}
\title{Annotate clusters with driver labels and typical-loss motifs}
\usage{
annotate_clusters(
  spectra,
  clusters,
  similarity,
  params = eihrms_default_params(),
  top_n_losses = 5,
  exclude_noise = TRUE,
  use_idf = TRUE
)
}
\arguments{
\item{spectra}{Result of build_spectra().}

\item{clusters}{Cluster vector (e.g., result of cluster_optics()$clusters). Names should be IDs.}

\item{similarity}{Result of compute_similarity_matrices().}

\item{params}{Parameter list.}

\item{top_n_losses}{Number of top typical losses to include as a semicolon-separated label.}

\item{exclude_noise}{Exclude cluster label "0".}

\item{use_idf}{Use IDF weighting for typical-loss motif ranking.}
}
\value{
A list with \code{membership}, \code{cluster_summary}, \code{typical_loss_long}.
}
\description{
This is a convenience wrapper that produces (i) membership table,
(ii) driver summary per cluster, and (iii) typical-loss motif tables.
}
