% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{apply_spectral_weights}
\alias{apply_spectral_weights}
\title{Apply Intensity and m/z Frequency Weighting to a Spectrum}
\usage{
apply_spectral_weights(
  spec,
  mz_freq = NULL,
  intensity_power = 0.5,
  freq_power = 0.5,
  bin_width = 1
)
}
\arguments{
\item{spec}{Spectrum matrix with columns (mz, intensity).}

\item{mz_freq}{Named m/z frequency vector from compute_mz_frequency().}

\item{intensity_power}{Power for intensity rescaling.}

\item{freq_power}{Power for m/z frequency weighting.}

\item{bin_width}{Bin width used to create mz_freq.}
}
\value{
Weighted spectrum matrix with columns (mz, intensity).
}
\description{
Apply Intensity and m/z Frequency Weighting to a Spectrum
}
