% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_msp.R
\name{build_spectra_from_msp}
\alias{build_spectra_from_msp}
\title{Build Spectra from MSP Library}
\usage{
build_spectra_from_msp(msp_file, params, require_ri = FALSE, progress = TRUE)
}
\arguments{
\item{msp_file}{Path to .msp file.}

\item{params}{Parameter list from eihrms_default_params().}

\item{require_ri}{If TRUE, filter to entries with RI values.}

\item{progress}{Show progress.}
}
\value{
A list similar to build_spectra() output.
}
\description{
Convenience function to process an MSP library through the full pipeline.
}
