% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_optics}
\alias{cluster_optics}
\title{Cluster with OPTICS on Distance Matrix}
\usage{
cluster_optics(sim_cluster, min_pts = NULL, eps_cl = 0.73)
}
\arguments{
\item{sim_cluster}{Similarity matrix for clustering.}

\item{min_pts}{Minimum points for OPTICS (defaults to max(3, floor(n/20))).}

\item{eps_cl}{Epsilon for extractDBSCAN.}
}
\value{
A list with optics result and cluster labels.
}
\description{
Cluster with OPTICS on Distance Matrix
}
