% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{compute_mz_frequency}
\alias{compute_mz_frequency}
\title{Compute m/z Frequency Map from Spectra}
\usage{
compute_mz_frequency(frag_list, bin_width = 1)
}
\arguments{
\item{frag_list}{Named list of fragment spectra matrices.}

\item{bin_width}{Bin width in Da for m/z grouping.}
}
\value{
Named numeric vector of m/z-bin frequencies in the range 0 to 1.
}
\description{
Compute m/z Frequency Map from Spectra
}
