% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_msp.R
\name{convert_msp_to_internal}
\alias{convert_msp_to_internal}
\title{Convert MSP Library to Package Format}
\usage{
convert_msp_to_internal(
  msp_lib,
  id_column = "name",
  require_ri = FALSE,
  require_inchikey = FALSE
)
}
\arguments{
\item{msp_lib}{Tibble from read_msp().}

\item{id_column}{Column to use as compound ID (default: "name").}

\item{require_ri}{If TRUE, filter to entries with RI values.}

\item{require_inchikey}{If TRUE, filter to entries with InChIKey.}
}
\value{
A tibble compatible with the package workflow.
}
\description{
Converts a tibble from read_msp() to the format used by build_spectra().
}
