% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{evaluate_precision_at_k}
\alias{evaluate_precision_at_k}
\title{Evaluate Precision at K}
\usage{
evaluate_precision_at_k(
  dist_mat,
  inchikeys,
  k_values = c(1, 5, 10, 20),
  inchikey_chars = 14
)
}
\arguments{
\item{dist_mat}{Distance matrix.}

\item{inchikeys}{Vector of InChIKeys for each spectrum.}

\item{k_values}{K values to evaluate.}

\item{inchikey_chars}{Number of InChIKey characters for grouping.}
}
\value{
List with precision_at_k, mean_average_precision, and n_queries.
}
\description{
Calculates mean precision at K and mean average precision for replicate queries.
}
