% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{evaluate_replicate_consistency}
\alias{evaluate_replicate_consistency}
\title{Evaluate Replicate Consistency}
\usage{
evaluate_replicate_consistency(dist_mat, inchikeys, ids)
}
\arguments{
\item{dist_mat}{Distance matrix.}

\item{inchikeys}{Vector of InChIKeys for each spectrum.}

\item{ids}{Vector of spectrum IDs.}
}
\value{
List with accuracy metrics.
}
\description{
Tests if spectra from the same compound (same InChIKey) are most similar.
}
