% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{evaluate_silhouette}
\alias{evaluate_silhouette}
\title{Evaluate Silhouette Score}
\usage{
evaluate_silhouette(dist_mat, labels)
}
\arguments{
\item{dist_mat}{Distance matrix.}

\item{labels}{Vector of class labels.}
}
\value{
List with silhouette metrics.
}
\description{
Calculates silhouette score using compound class as labels.
}
