% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{generate_demo_msdial}
\alias{generate_demo_msdial}
\title{Generate Synthetic MS-DIAL-like Demo Data}
\usage{
generate_demo_msdial(
  n_known = 5,
  n_unknown = 5,
  n_samples = 6,
  n_blanks = 1,
  n_standards = 1,
  seed = 1
)
}
\arguments{
\item{n_known}{Number of known compounds.}

\item{n_unknown}{Number of unknown compounds.}

\item{n_samples}{Number of sample columns.}

\item{n_blanks}{Number of blank columns.}

\item{n_standards}{Number of standard columns.}

\item{seed}{Random seed for reproducibility.}
}
\value{
A tibble resembling an MS-DIAL export table.
}
\description{
Generate Synthetic MS-DIAL-like Demo Data
}
