% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_derivatization.R
\name{roc_auc}
\alias{roc_auc}
\title{AUC (trapezoid) from an ROC curve}
\usage{
roc_auc(roc_df)
}
\arguments{
\item{roc_df}{Output of roc_curve_binary().}
}
\value{
Numeric AUC.
}
\description{
AUC (trapezoid) from an ROC curve
}
