% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_msp.R
\name{write_msp}
\alias{write_msp}
\title{Write Spectra to MSP Format}
\usage{
write_msp(df_spec, frag_list, file, include_losses = FALSE)
}
\arguments{
\item{df_spec}{Data frame with spectral information.}

\item{frag_list}{Named list of fragment spectra matrices.}

\item{file}{Output file path.}

\item{include_losses}{If TRUE, also write loss spectra.}
}
\description{
Export spectra to MSP format for sharing or use in other software.
}
